+++ hds-server/src/main/java/com/tecnico/sec/hds/server/db/rules/ReceiveAmountRules.java
@@ -21,7 +21,7 @@ public class ReceiveAmountRules {
       Optional<Transaction> transaction = transferQueries.getTransactionByHash(transHash);
 
       if (transaction.isPresent() && transaction.get().pending) {
-        float amount = transaction.get().amount;
+        long amount = transaction.get().amount;
         String sourceKey = transaction.get().sourceKey;
         String destKey = transaction.get().destKey;
 
@@ -36,7 +36,7 @@ public class ReceiveAmountRules {
             ChainHelper.TransactionType.ACCEPT,
             signature);
 
-        float balance = accountQueries.getBalance(destKey);
+        long balance = accountQueries.getBalance(destKey);
 
         transferQueries.updateTransactionPendingState(transHash, false);
         transferQueries.insertNewTransaction(sourceKey, destKey, amount, false, true, signature, newHash);
